% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanRBC.function.R
\name{meanRBC.function}
\alias{meanRBC.function}
\title{Mean red blood cell age}
\usage{
meanRBC.function(time, APE)
}
\arguments{
\item{time}{A vector of time points at which the relative APE is collected}

\item{APE}{A vector of relative APE for each time measurement}
}
\value{
The mean red blood cell age over the time measured
}
\description{
Estimates the mean red blood cell age from an estimated survival curve
}
\examples{
time <- c(0,1,7,14,21,28,42,56,70,84,98,105,112,119,126,140,161,182); 
APE <- c(NA,2.3, 56.62, 87.15, 86.56, 90.60, 98.46, 89.64, 100, 81.19, 86.95,83.69, 83.88, 63.53, 64.30, 33.97, 22.84, 18.43);
mrbc <- meanRBC.function(time, APE)
}
